add custom block variation __experimentalConfirmDialogの方がシンプルになるので修正

+++ src/extensions/common/custom-block-variation/block-variation-list/item/title-area/delete-button/index.js
@@ -3,7 +3,10 @@
  */
 import { __ } from '@wordpress/i18n';
 import { useState } from '@wordpress/element';
-import { Button, Modal, Flex, FlexItem } from '@wordpress/components';
+import {
+	Button,
+	__experimentalConfirmDialog as ConfirmDialog,
+} from '@wordpress/components';
 
 export const DeleteButton = ({ index, variationState, setVariationState }) => {
 	const [isModalOpen, setIsModalOpen] = useState(false);
@@ -24,41 +27,24 @@ export const DeleteButton = ({ index, variationState, setVariationState }) => {
 			>
 				{__('Delete', 'vk-blocks-pro')}
 			</Button>
-			{isModalOpen && (
-				<Modal
-					title={__('Would you like to delete?', 'vk-blocks-pro')}
-					overlayClassName="custom-block-variation__confirmation-modal"
-					onRequestClose={closeModal}
-					shouldCloseOnClickOutside={false}
-				>
-					<div className="custom_block_style_delete_modal">
-						<div className="custom_block_style_delete_modal_button_area">
-							<Flex justify="flex-end">
-								<FlexItem>
-									<Button
-										variant="secondary"
-										onClick={closeModal}
-									>
-										{__('Cancel')}
-									</Button>
-								</FlexItem>
-								<FlexItem>
-									<Button
-										className="delete-item-button"
-										isDestructive
-										onClick={() => {
-											deleteItem();
-											closeModal();
-										}}
-									>
-										{__('Delete', 'vk-blocks-pro')}
-									</Button>
-								</FlexItem>
-							</Flex>
-						</div>
-					</div>
-				</Modal>
-			)}
+			<ConfirmDialog
+				isOpen={isModalOpen}
+				cancelButtonText={__('Cancel')}
+				confirmButtonText={__('Delete', 'vk-blocks-pro')}
+				onConfirm={() => {
+					deleteItem();
+					closeModal();
+				}}
+				onCancel={() => {
+					closeModal();
+				}}
+			>
+				{__(
+					'このバリエーションを削除してもよろしいですか?',
+					// 'Are you sure you want to delete this variation?',
+					'vk-blocks-pro'
+				)}
+			</ConfirmDialog>
 		</div>
 	);
 };
